Naming refactoring
