+++ src/components/Cities/Cities.tsx
@@ -1,6 +1,6 @@
 import {useState} from 'react';
-import {PlaceCard} from '../place-card/place-card.tsx';
-import {Map} from '../map/map.tsx';
+import {PlaceCard} from '../PlaceCard/PlaceCard.tsx';
+import {Map} from '../Map/Map.tsx';
 import {TOffers} from '../../types/offer.ts';
 
 
@@ -8,7 +8,7 @@ type TCitiesProps = {
   offers: TOffers;
 }
 
-function Cities({offers}: TCitiesProps) {
+const Cities = ({offers}: TCitiesProps) => {
   const [, setActiveOffer] = useState(null);
 
   const handleCardMouseMove = (id: number) => {
@@ -68,6 +68,6 @@ function Cities({offers}: TCitiesProps) {
       </div>
     </div>
   );
-}
+};
 
 export {Cities};
