+++ src/components/PrivateRoute/PrivateRoute.tsx
@@ -1,12 +1,13 @@
 import {Navigate} from 'react-router-dom';
 import {AppRoute, AuthorizationStatus} from '../../const.ts';
 
+
 type TPrivateRouteProps = {
   authorizationStatus: AuthorizationStatus;
   children: JSX.Element;
 }
 
-function PrivateRoute(props: TPrivateRouteProps) {
+const PrivateRoute = (props: TPrivateRouteProps) => {
   const {authorizationStatus, children} = props;
 
   return (
@@ -14,6 +15,6 @@ function PrivateRoute(props: TPrivateRouteProps) {
       ? children
       : <Navigate to={AppRoute.Login}/>
   );
-}
+};
 
 export {PrivateRoute};
