+++ src/components/ReviewForm/ReviewForm.tsx
@@ -3,7 +3,7 @@ import type {ChangeEvent} from 'react';
 import {STARS_COUNT} from '../../const.ts';
 
 
-function ReviewForm() {
+const ReviewForm = () => {
   const [comment, setComment] = useState('');
   const [rating, setRating] = useState('');
   const isValid =
@@ -72,6 +72,6 @@ function ReviewForm() {
       </div>
     </form>
   );
-}
+};
 
 export {ReviewForm};
