+++ src/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import {App} from './components/app/app.tsx';
+import {App} from './App.tsx';
 import {offers} from './mocks/offers.ts';
 import {comments} from './mocks/comments.ts';
